refactor API interaction into a nice little module
should make it easier to expand the api surface as tb evolves too
