+++ src/api/debug.ts
@@ -1,0 +1,4 @@
+import debugFactory from "../debug";
+const debug: debug.IDebugger = debugFactory.extend("api");
+
+export default debug;
