+++ src/api/index.ts
@@ -1,0 +1,10 @@
+import { AuthError, APIError } from "./errors";
+
+import Bookmarks from "./resources/bookmarks";
+
+const API = {
+  Bookmarks
+};
+
+export default API;
+export { AuthError, APIError };
