+++ src/api/types.d.ts
@@ -1,0 +1,8 @@
+export interface Bookmark {
+  id?: string | number;
+  title: string;
+  url: string;
+  tags: string[];
+  description: string;
+  [keys: string]: any;
+}
