+++ src/debug.tsx
@@ -12,6 +12,11 @@ export type DebugData = {
 
 export const debugInstances: DebugData[] = [
   {
+    namespace: "API Requests",
+    description: ``,
+    debugFlag: "transientBug:api*"
+  },
+  {
     namespace: "Background Page",
     description: `Includes login and background saving functionality along with post-install setup.`,
     debugFlag: "transientBug:pages:Background*"
@@ -31,7 +36,7 @@ export const debugInstances: DebugData[] = [
     description: (
       <p>
         Enables <b>ALL</b> debug logging in the extension. This can be noisy and
-        should be used as a last resort.
+        should be used as a last resort or in development settings.
       </p>
     ),
     debugFlag: "transientBug*"
